Add subscribers for lifecycle events. (#10)
* Add subscribers for lifecycle events.

There is a basic example of configuration, currently in subscribers.py,
this may or may not be the best place for it. A substantial doc
reorginazation can come after #8, #9 and issue #6 (PR #10, this one) are
merged.

Fixes #6

* typos

* Refactor tests.

Separate each main functional test into its own file. This greatly
simplifies the imports of most of them.

Also simplify the subscriber tests by removing some partially duplicated
code.

* Correct wrong interface name in subscribers.zcml

Test this file by relying on it to configure the subscribers in
test_subscribers.py.